Split instance service into an explicit and a dummy implementation. Default the dummy will be provided, to not cause errors.
